Return reference to VfbConnect object

Return reference to VfbConnect object

Usage

VfbConnect(..., raw = FALSE, cache = TRUE)

Arguments

...

Additional arguments passed to Python's `VfbConnect`. See examples.

raw

Whether to return python list containing raw JSON (default FALSE)

cache

Whether to cache the VfbConnect object (default T since this can save some time).

Value

An list containing the processed results or (when raw=TRUE) as python list containing raw JSON.

Details

See https://pypi.org/project/vfb-connect/. The raw=TRUE option is mostly useful to see what output would look like.

Examples

vc=VfbConnect()
# large return, so use str
str(vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686')), max=2)
#> List of 1
#>  $ :List of 10
#>   ..$ term                 :List of 3
#>   ..$ query                : chr "Get JSON for Class"
#>   ..$ version              : chr "c58c844"
#>   ..$ parents              :List of 2
#>   ..$ relationships        :List of 3
#>   ..$ related_individuals  : list()
#>   ..$ xrefs                : list()
#>   ..$ anatomy_channel_image:List of 10
#>   ..$ pub_syn              :List of 6
#>   ..$ def_pubs             :List of 3
# do this to explore
# vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686')) %>% View()
vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE)
#> [[1]]
#> [[1]]$label
#> [1] "Kenyon cell"
#> 
#> [[1]]$symbol
#> [1] "KC"
#> 
#> [[1]]$id
#> [1] "FBbt_00003686"
#> 
#> [[1]]$tags
#> [1] "Entity|Anatomy|Cell|Class|Nervous_system|Neuron|has_subClass|hasScRNAseq"
#> 
#> [[1]]$parents_label
#> [1] "anterior ectoderm derivative|mushroom body intrinsic neuron"
#> 
#> [[1]]$parents_id
#> [1] "FBbt_00025991|FBbt_00007484"
#> 
#> 

vc$neo_query_wrapper$get_DataSet_TermInfo(list('Ito2013'), summary=TRUE)
#> [[1]]
#> [[1]]$label
#> [1] "Ito lab adult brain lineage clone image set"
#> 
#> [[1]]$symbol
#> [1] ""
#> 
#> [[1]]$id
#> [1] "Ito2013"
#> 
#> [[1]]$tags
#> [1] "Entity|DataSet|Individual|has_image"
#> 
#> [[1]]$description
#> [1] "An exhaustive set of lineage clones covering the adult brain from Kei Ito's  lab."
#> 
#> [[1]]$miniref
#> [1] "Ito et al., 2013, Curr. Biol. 23(8): 644--655"
#> 
#> [[1]]$FlyBase
#> [1] "FBrf0221438"
#> 
#> [[1]]$PMID
#> [1] "23541729"
#> 
#> [[1]]$DOI
#> [1] "10.1016/j.cub.2013.03.015"
#> 
#> 

vc$neo_query_wrapper$get_anatomical_individual_TermInfo(list('VFB_00010001'))
#> [[1]]
#> [[1]]$term
#> [[1]]$term$core
#> [[1]]$term$core$iri
#> [1] "http://virtualflybrain.org/reports/VFB_00010001"
#> 
#> [[1]]$term$core$symbol
#> [1] ""
#> 
#> [[1]]$term$core$types
#>  [1] "Entity"                      "Adult"                      
#>  [3] "Anatomy"                     "Cell"                       
#>  [5] "Expression_pattern_fragment" "Individual"                 
#>  [7] "Nervous_system"              "Neuron"                     
#>  [9] "VFB"                         "has_image"                  
#> [11] "FlyCircuit"                  "NBLAST"                     
#> 
#> [[1]]$term$core$short_form
#> [1] "VFB_00010001"
#> 
#> [[1]]$term$core$unique_facets
#> [1] "Adult"                       "Expression_pattern_fragment"
#> [3] "Nervous_system"              "Neuron"                     
#> 
#> [[1]]$term$core$label
#> [1] "fru-F-500075"
#> 
#> 
#> [[1]]$term$description
#> list()
#> 
#> [[1]]$term$comment
#> [1] "OutAge: Adult 5~15 days"
#> 
#> 
#> [[1]]$query
#> [1] "Get JSON for Individual"
#> 
#> [[1]]$version
#> [1] "c58c844"
#> 
#> [[1]]$dataset_license
#> [[1]]$dataset_license[[1]]
#> [[1]]$dataset_license[[1]]$license
#> [[1]]$dataset_license[[1]]$license$core
#> [[1]]$dataset_license[[1]]$license$core$iri
#> [1] "http://virtualflybrain.org/reports/VFBlicense_FlyCircuit_License"
#> 
#> [[1]]$dataset_license[[1]]$license$core$symbol
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$license$core$types
#> [1] "Entity"     "Individual" "License"   
#> 
#> [[1]]$dataset_license[[1]]$license$core$short_form
#> [1] "VFBlicense_FlyCircuit_License"
#> 
#> [[1]]$dataset_license[[1]]$license$core$unique_facets
#> [1] "License"
#> 
#> [[1]]$dataset_license[[1]]$license$core$label
#> [1] "FlyCircuit License"
#> 
#> 
#> [[1]]$dataset_license[[1]]$license$icon
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$license$link
#> [1] ""
#> 
#> 
#> [[1]]$dataset_license[[1]]$dataset
#> [[1]]$dataset_license[[1]]$dataset$core
#> [[1]]$dataset_license[[1]]$dataset$core$iri
#> [1] "http://virtualflybrain.org/reports/Chiang2010"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$symbol
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$types
#> [1] "Entity"     "DataSet"    "Individual" "has_image"  "FlyCircuit"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$short_form
#> [1] "Chiang2010"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$unique_facets
#> [1] "DataSet"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$label
#> [1] "FlyCircuit 1.0 - single neurons (Chiang2010)"
#> 
#> 
#> [[1]]$dataset_license[[1]]$dataset$link
#> [1] "http://flybase.org/reports/FBrf0210580.html"
#> 
#> 
#> 
#> 
#> [[1]]$parents
#> [[1]]$parents[[1]]
#> [[1]]$parents[[1]]$symbol
#> [1] ""
#> 
#> [[1]]$parents[[1]]$iri
#> [1] "http://purl.obolibrary.org/obo/fbbt/vfb/VFBext_0000004"
#> 
#> [[1]]$parents[[1]]$types
#> [1] "Entity"                      "Anatomy"                    
#> [3] "Class"                       "Expression_pattern_fragment"
#> 
#> [[1]]$parents[[1]]$short_form
#> [1] "VFBext_0000004"
#> 
#> [[1]]$parents[[1]]$unique_facets
#> [1] "Anatomy"
#> 
#> [[1]]$parents[[1]]$label
#> [1] "expression pattern fragment"
#> 
#> 
#> [[1]]$parents[[2]]
#> [[1]]$parents[[2]]$symbol
#> [1] ""
#> 
#> [[1]]$parents[[2]]$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00050031"
#> 
#> [[1]]$parents[[2]]$types
#> [1] "Entity"         "Adult"          "Anatomy"        "Cell"          
#> [5] "Class"          "Nervous_system" "Neuron"         "has_subClass"  
#> 
#> [[1]]$parents[[2]]$short_form
#> [1] "FBbt_00050031"
#> 
#> [[1]]$parents[[2]]$unique_facets
#> [1] "Adult"          "Nervous_system" "Neuron"        
#> 
#> [[1]]$parents[[2]]$label
#> [1] "adult SMPpv1 lineage neuron"
#> 
#> 
#> 
#> [[1]]$relationships
#> [[1]]$relationships[[1]]
#> [[1]]$relationships[[1]]$relation
#> [[1]]$relationships[[1]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/RO_0002292"
#> 
#> [[1]]$relationships[[1]]$relation$database_cross_reference
#> list()
#> 
#> [[1]]$relationships[[1]]$relation$label
#> [1] "expresses"
#> 
#> [[1]]$relationships[[1]]$relation$type
#> [1] "expresses"
#> 
#> [[1]]$relationships[[1]]$relation$confidence_value
#> [1] ""
#> 
#> 
#> [[1]]$relationships[[1]]$object
#> [[1]]$relationships[[1]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[1]]$object$iri
#> [1] "http://flybase.org/reports/FBal0276838"
#> 
#> [[1]]$relationships[[1]]$object$types
#> [1] "Entity"  "Class"   "Feature"
#> 
#> [[1]]$relationships[[1]]$object$short_form
#> [1] "FBal0276838"
#> 
#> [[1]]$relationships[[1]]$object$unique_facets
#> [1] "Allele"
#> 
#> [[1]]$relationships[[1]]$object$label
#> [1] "Scer\\GAL4[fru.P1.D]"
#> 
#> 
#> 
#> [[1]]$relationships[[2]]
#> [[1]]$relationships[[2]]$relation
#> [[1]]$relationships[[2]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/RO_0002131"
#> 
#> [[1]]$relationships[[2]]$relation$database_cross_reference
#> list()
#> 
#> [[1]]$relationships[[2]]$relation$label
#> [1] "overlaps"
#> 
#> [[1]]$relationships[[2]]$relation$type
#> [1] "overlaps"
#> 
#> [[1]]$relationships[[2]]$relation$confidence_value
#> [1] ""
#> 
#> 
#> [[1]]$relationships[[2]]$object
#> [[1]]$relationships[[2]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[2]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00045050"
#> 
#> [[1]]$relationships[[2]]$object$types
#> [1] "Entity"                   "Adult"                   
#> [3] "Anatomy"                  "Class"                   
#> [5] "Nervous_system"           "Synaptic_neuropil"       
#> [7] "Synaptic_neuropil_domain"
#> 
#> [[1]]$relationships[[2]]$object$short_form
#> [1] "FBbt_00045050"
#> 
#> [[1]]$relationships[[2]]$object$unique_facets
#> [1] "Adult"                    "Nervous_system"          
#> [3] "Synaptic_neuropil_domain"
#> 
#> [[1]]$relationships[[2]]$object$label
#> [1] "flange"
#> 
#> 
#> 
#> [[1]]$relationships[[3]]
#> [[1]]$relationships[[3]]$relation
#> [[1]]$relationships[[3]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/RO_0002131"
#> 
#> [[1]]$relationships[[3]]$relation$database_cross_reference
#> list()
#> 
#> [[1]]$relationships[[3]]$relation$label
#> [1] "overlaps"
#> 
#> [[1]]$relationships[[3]]$relation$type
#> [1] "overlaps"
#> 
#> [[1]]$relationships[[3]]$relation$confidence_value
#> [1] ""
#> 
#> 
#> [[1]]$relationships[[3]]$object
#> [[1]]$relationships[[3]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[3]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00007055"
#> 
#> [[1]]$relationships[[3]]$object$types
#> [1] "Entity"                   "Adult"                   
#> [3] "Anatomy"                  "Class"                   
#> [5] "Nervous_system"           "Synaptic_neuropil"       
#> [7] "Synaptic_neuropil_domain"
#> 
#> [[1]]$relationships[[3]]$object$short_form
#> [1] "FBbt_00007055"
#> 
#> [[1]]$relationships[[3]]$object$unique_facets
#> [1] "Adult"                    "Nervous_system"          
#> [3] "Synaptic_neuropil_domain"
#> 
#> [[1]]$relationships[[3]]$object$label
#> [1] "superior medial protocerebrum"
#> 
#> 
#> 
#> [[1]]$relationships[[4]]
#> [[1]]$relationships[[4]]$relation
#> [[1]]$relationships[[4]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/BFO_0000050"
#> 
#> [[1]]$relationships[[4]]$relation$database_cross_reference
#> list()
#> 
#> [[1]]$relationships[[4]]$relation$label
#> [1] "is part of"
#> 
#> [[1]]$relationships[[4]]$relation$type
#> [1] "part_of"
#> 
#> [[1]]$relationships[[4]]$relation$confidence_value
#> [1] ""
#> 
#> 
#> [[1]]$relationships[[4]]$object
#> [[1]]$relationships[[4]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[4]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00003624"
#> 
#> [[1]]$relationships[[4]]$object$types
#> [1] "Entity"         "Adult"          "Anatomy"        "Class"         
#> [5] "Nervous_system"
#> 
#> [[1]]$relationships[[4]]$object$short_form
#> [1] "FBbt_00003624"
#> 
#> [[1]]$relationships[[4]]$object$unique_facets
#> [1] "Adult"          "Nervous_system"
#> 
#> [[1]]$relationships[[4]]$object$label
#> [1] "adult brain"
#> 
#> 
#> 
#> [[1]]$relationships[[5]]
#> [[1]]$relationships[[5]]$relation
#> [[1]]$relationships[[5]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/BFO_0000050"
#> 
#> [[1]]$relationships[[5]]$relation$database_cross_reference
#> list()
#> 
#> [[1]]$relationships[[5]]$relation$label
#> [1] "is part of"
#> 
#> [[1]]$relationships[[5]]$relation$type
#> [1] "part_of"
#> 
#> [[1]]$relationships[[5]]$relation$confidence_value
#> [1] ""
#> 
#> 
#> [[1]]$relationships[[5]]$object
#> [[1]]$relationships[[5]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[5]]$object$iri
#> [1] "http://virtualflybrain.org/reports/VFBexp_FBal0276838"
#> 
#> [[1]]$relationships[[5]]$object$types
#> [1] "Entity"             "Anatomy"            "Class"             
#> [4] "Expression_pattern"
#> 
#> [[1]]$relationships[[5]]$object$short_form
#> [1] "VFBexp_FBal0276838"
#> 
#> [[1]]$relationships[[5]]$object$unique_facets
#> [1] "Expression_pattern"
#> 
#> [[1]]$relationships[[5]]$object$label
#> [1] "Scer\\GAL4[fru.P1.D] expression pattern"
#> 
#> 
#> 
#> [[1]]$relationships[[6]]
#> [[1]]$relationships[[6]]$relation
#> [[1]]$relationships[[6]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/BFO_0000050"
#> 
#> [[1]]$relationships[[6]]$relation$database_cross_reference
#> list()
#> 
#> [[1]]$relationships[[6]]$relation$label
#> [1] "is part of"
#> 
#> [[1]]$relationships[[6]]$relation$type
#> [1] "part_of"
#> 
#> [[1]]$relationships[[6]]$relation$confidence_value
#> [1] ""
#> 
#> 
#> [[1]]$relationships[[6]]$object
#> [[1]]$relationships[[6]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[6]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00007011"
#> 
#> [[1]]$relationships[[6]]$object$types
#> [1] "Entity"  "Anatomy" "Class"  
#> 
#> [[1]]$relationships[[6]]$object$short_form
#> [1] "FBbt_00007011"
#> 
#> [[1]]$relationships[[6]]$object$unique_facets
#> [1] "Anatomy"
#> 
#> [[1]]$relationships[[6]]$object$label
#> [1] "female organism"
#> 
#> 
#> 
#> 
#> [[1]]$related_individuals
#> list()
#> 
#> [[1]]$xrefs
#> [[1]]$xrefs[[1]]
#> [[1]]$xrefs[[1]]$link_base
#> [1] "http://flycircuit.tw/modules.php?name=clearpage&op=detail_table&neuron="
#> 
#> [[1]]$xrefs[[1]]$site
#> [[1]]$xrefs[[1]]$site$symbol
#> [1] ""
#> 
#> [[1]]$xrefs[[1]]$site$iri
#> [1] "http://virtualflybrain.org/reports/FlyCircuit"
#> 
#> [[1]]$xrefs[[1]]$site$types
#> [1] "Entity"     "Individual" "Site"      
#> 
#> [[1]]$xrefs[[1]]$site$short_form
#> [1] "FlyCircuit"
#> 
#> [[1]]$xrefs[[1]]$site$unique_facets
#> [1] "Site"
#> 
#> [[1]]$xrefs[[1]]$site$label
#> [1] "FlyCircuit 1.0"
#> 
#> 
#> [[1]]$xrefs[[1]]$is_data_source
#> [1] TRUE
#> 
#> [[1]]$xrefs[[1]]$link_postfix
#> [1] ""
#> 
#> [[1]]$xrefs[[1]]$icon
#> [1] ""
#> 
#> [[1]]$xrefs[[1]]$accession
#> [1] "fru-F-500075"
#> 
#> [[1]]$xrefs[[1]]$link_text
#> [1] "fru-F-500075 on FlyCircuit 1.0"
#> 
#> [[1]]$xrefs[[1]]$homepage
#> [1] "http://flycircuit.tw"
#> 
#> 
#> 
#> [[1]]$channel_image
#> [[1]]$channel_image[[1]]
#> [[1]]$channel_image[[1]]$image
#> [[1]]$channel_image[[1]]$image$image_nrrd
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/volume.nrrd"
#> 
#> [[1]]$channel_image[[1]]$image$image_swc
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/volume.swc"
#> 
#> [[1]]$channel_image[[1]]$image$template_channel
#> [[1]]$channel_image[[1]]$image$template_channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00101567"
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$types
#> [1] "Entity"     "Individual" "Template"  
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$short_form
#> [1] "VFBc_00101567"
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$unique_facets
#> list()
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$label
#> [1] "JRC2018Unisex_c"
#> 
#> 
#> [[1]]$channel_image[[1]]$image$index
#> list()
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy
#> [[1]]$channel_image[[1]]$image$template_anatomy$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$iri
#> [1] "http://virtualflybrain.org/reports/VFB_00101567"
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$types
#> [1] "Entity"         "Adult"          "Anatomy"        "Individual"    
#> [5] "Nervous_system" "Template"       "has_image"     
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$short_form
#> [1] "VFB_00101567"
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$unique_facets
#> [1] "Adult"          "Nervous_system"
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$label
#> [1] "JRC2018Unisex"
#> 
#> 
#> [[1]]$channel_image[[1]]$image$image_wlz
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/volume.wlz"
#> 
#> [[1]]$channel_image[[1]]$image$image_obj
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/volume.obj"
#> 
#> [[1]]$channel_image[[1]]$image$image_thumbnail
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/thumbnail.png"
#> 
#> [[1]]$channel_image[[1]]$image$image_folder
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/"
#> 
#> 
#> [[1]]$channel_image[[1]]$channel
#> [[1]]$channel_image[[1]]$channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[1]]$channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00010001"
#> 
#> [[1]]$channel_image[[1]]$channel$types
#> [1] "Entity"     "Individual" "VFB"       
#> 
#> [[1]]$channel_image[[1]]$channel$short_form
#> [1] "VFBc_00010001"
#> 
#> [[1]]$channel_image[[1]]$channel$unique_facets
#> list()
#> 
#> [[1]]$channel_image[[1]]$channel$label
#> [1] "fru-F-500075-c"
#> 
#> 
#> [[1]]$channel_image[[1]]$imaging_technique
#> [[1]]$channel_image[[1]]$imaging_technique$symbol
#> [1] "Confocal"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$iri
#> [1] "http://purl.obolibrary.org/obo/FBbi_00000251"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$types
#> [1] "Entity"       "Class"        "has_subClass"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$short_form
#> [1] "FBbi_00000251"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$unique_facets
#> [1] "Class"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$label
#> [1] "confocal microscopy"
#> 
#> 
#> 
#> [[1]]$channel_image[[2]]
#> [[1]]$channel_image[[2]]$image
#> [[1]]$channel_image[[2]]$image$image_nrrd
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00017894/volume.nrrd"
#> 
#> [[1]]$channel_image[[2]]$image$image_swc
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00017894/volume.swc"
#> 
#> [[1]]$channel_image[[2]]$image$template_channel
#> [[1]]$channel_image[[2]]$image$template_channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00017894"
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$types
#> [1] "Entity"     "Individual" "Template"   "VFB"       
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$short_form
#> [1] "VFBc_00017894"
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$unique_facets
#> list()
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$label
#> [1] "JFRC2_template_c"
#> 
#> 
#> [[1]]$channel_image[[2]]$image$index
#> list()
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy
#> [[1]]$channel_image[[2]]$image$template_anatomy$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$iri
#> [1] "http://virtualflybrain.org/reports/VFB_00017894"
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$types
#> [1] "Entity"         "Adult"          "Anatomy"        "Individual"    
#> [5] "Nervous_system" "Template"       "VFB"            "has_image"     
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$short_form
#> [1] "VFB_00017894"
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$unique_facets
#> [1] "Adult"          "Nervous_system"
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$label
#> [1] "adult brain template JFRC2"
#> 
#> 
#> [[1]]$channel_image[[2]]$image$image_wlz
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00017894/volume.wlz"
#> 
#> [[1]]$channel_image[[2]]$image$image_obj
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00017894/volume.obj"
#> 
#> [[1]]$channel_image[[2]]$image$image_thumbnail
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00017894/thumbnail.png"
#> 
#> [[1]]$channel_image[[2]]$image$image_folder
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00017894/"
#> 
#> 
#> [[1]]$channel_image[[2]]$channel
#> [[1]]$channel_image[[2]]$channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[2]]$channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00010001"
#> 
#> [[1]]$channel_image[[2]]$channel$types
#> [1] "Entity"     "Individual" "VFB"       
#> 
#> [[1]]$channel_image[[2]]$channel$short_form
#> [1] "VFBc_00010001"
#> 
#> [[1]]$channel_image[[2]]$channel$unique_facets
#> list()
#> 
#> [[1]]$channel_image[[2]]$channel$label
#> [1] "fru-F-500075-c"
#> 
#> 
#> [[1]]$channel_image[[2]]$imaging_technique
#> [[1]]$channel_image[[2]]$imaging_technique$symbol
#> [1] "Confocal"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$iri
#> [1] "http://purl.obolibrary.org/obo/FBbi_00000251"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$types
#> [1] "Entity"       "Class"        "has_subClass"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$short_form
#> [1] "FBbi_00000251"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$unique_facets
#> [1] "Class"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$label
#> [1] "confocal microscopy"
#> 
#> 
#> 
#> 
#> [[1]]$pub_syn
#> [[1]]$pub_syn[[1]]
#> [[1]]$pub_syn[[1]]$synonym
#> [[1]]$pub_syn[[1]]$synonym$scope
#> [1] "has_exact_synonym"
#> 
#> [[1]]$pub_syn[[1]]$synonym$label
#> [1] "FruMARCM-F000584_seg002"
#> 
#> [[1]]$pub_syn[[1]]$synonym$type
#> [1] ""
#> 
#> 
#> [[1]]$pub_syn[[1]]$pub
#> [[1]]$pub_syn[[1]]$pub$core
#> [[1]]$pub_syn[[1]]$pub$core$symbol
#> [1] ""
#> 
#> [[1]]$pub_syn[[1]]$pub$core$iri
#> [1] "http://flybase.org/reports/Unattributed"
#> 
#> [[1]]$pub_syn[[1]]$pub$core$types
#> [1] "Entity"     "Individual" "pub"       
#> 
#> [[1]]$pub_syn[[1]]$pub$core$short_form
#> [1] "Unattributed"
#> 
#> [[1]]$pub_syn[[1]]$pub$core$unique_facets
#> [1] "pub"
#> 
#> [[1]]$pub_syn[[1]]$pub$core$label
#> [1] ""
#> 
#> 
#> [[1]]$pub_syn[[1]]$pub$FlyBase
#> [1] ""
#> 
#> [[1]]$pub_syn[[1]]$pub$PubMed
#> [1] ""
#> 
#> [[1]]$pub_syn[[1]]$pub$DOI
#> [1] ""
#> 
#> 
#> 
#> [[1]]$pub_syn[[2]]
#> [[1]]$pub_syn[[2]]$synonym
#> [[1]]$pub_syn[[2]]$synonym$scope
#> [1] "has_exact_synonym"
#> 
#> [[1]]$pub_syn[[2]]$synonym$label
#> [1] "fru-F-500075"
#> 
#> [[1]]$pub_syn[[2]]$synonym$type
#> [1] ""
#> 
#> 
#> [[1]]$pub_syn[[2]]$pub
#> [[1]]$pub_syn[[2]]$pub$core
#> [[1]]$pub_syn[[2]]$pub$core$symbol
#> [1] ""
#> 
#> [[1]]$pub_syn[[2]]$pub$core$iri
#> [1] "http://flybase.org/reports/Unattributed"
#> 
#> [[1]]$pub_syn[[2]]$pub$core$types
#> [1] "Entity"     "Individual" "pub"       
#> 
#> [[1]]$pub_syn[[2]]$pub$core$short_form
#> [1] "Unattributed"
#> 
#> [[1]]$pub_syn[[2]]$pub$core$unique_facets
#> [1] "pub"
#> 
#> [[1]]$pub_syn[[2]]$pub$core$label
#> [1] ""
#> 
#> 
#> [[1]]$pub_syn[[2]]$pub$FlyBase
#> [1] ""
#> 
#> [[1]]$pub_syn[[2]]$pub$PubMed
#> [1] ""
#> 
#> [[1]]$pub_syn[[2]]$pub$DOI
#> [1] ""
#> 
#> 
#> 
#> 
#> [[1]]$def_pubs
#> list()
#> 
#> 
# Get all terms relevant to a brain region (all parts and all overlapping cells$
# Query by label supported by default
vc$get_terms_by_region('fan-shaped body') %>%
  vc_df() %>%
  head(10)
#> Error in (function (..., row.names = NULL, check.rows = FALSE, check.names = TRUE,     fix.empty.names = TRUE, stringsAsFactors = FALSE) {    data.row.names <- if (check.rows && is.null(row.names))         function(current, new, i) {            if (is.character(current))                 new <- as.character(new)            if (is.character(new))                 current <- as.character(current)            if (anyDuplicated(new))                 return(current)            if (is.null(current))                 return(new)            if (all(current == new) || all(current == ""))                 return(new)            stop(gettextf("mismatch of row names in arguments of 'data.frame', item %d",                 i), domain = NA)        }    else function(current, new, i) {        current %||% if (anyDuplicated(new)) {            warning(gettextf("some row.names duplicated: %s --> row.names NOT used",                 paste(which(duplicated(new)), collapse = ",")),                 domain = NA)            current        }        else new    }    object <- as.list(substitute(list(...)))[-1L]    mirn <- missing(row.names)    mrn <- is.null(row.names)    x <- list(...)    n <- length(x)    if (n < 1L) {        if (!mrn) {            if (is.object(row.names) || !is.integer(row.names))                 row.names <- as.character(row.names)            if (anyNA(row.names))                 stop("row names contain missing values")            if (anyDuplicated(row.names))                 stop(gettextf("duplicate row.names: %s", paste(unique(row.names[duplicated(row.names)]),                   collapse = ", ")), domain = NA)        }        else row.names <- integer()        return(structure(list(), names = character(), row.names = row.names,             class = "data.frame"))    }    vnames <- names(x)    if (length(vnames) != n)         vnames <- character(n)    no.vn <- !nzchar(vnames)    vlist <- vnames <- as.list(vnames)    nrows <- ncols <- integer(n)    for (i in seq_len(n)) {        xi <- if (is.character(x[[i]]) || is.list(x[[i]]))             as.data.frame(x[[i]], optional = TRUE, stringsAsFactors = stringsAsFactors)        else as.data.frame(x[[i]], optional = TRUE)        nrows[i] <- .row_names_info(xi)        ncols[i] <- length(xi)        namesi <- names(xi)        if (ncols[i] > 1L) {            if (length(namesi) == 0L)                 namesi <- seq_len(ncols[i])            vnames[[i]] <- if (no.vn[i])                 namesi            else paste(vnames[[i]], namesi, sep = ".")        }        else if (length(namesi)) {            vnames[[i]] <- namesi        }        else if (fix.empty.names && no.vn[[i]]) {            tmpname <- deparse(object[[i]], nlines = 1L)[1L]            if (startsWith(tmpname, "I(") && endsWith(tmpname,                 ")")) {                ntmpn <- nchar(tmpname, "c")                tmpname <- substr(tmpname, 3L, ntmpn - 1L)            }            vnames[[i]] <- tmpname        }        if (mirn && nrows[i] > 0L) {            rowsi <- attr(xi, "row.names")            if (any(nzchar(rowsi)))                 row.names <- data.row.names(row.names, rowsi,                   i)        }        nrows[i] <- abs(nrows[i])        vlist[[i]] <- xi    }    nr <- max(nrows)    for (i in seq_len(n)[nrows < nr]) {        xi <- vlist[[i]]        if (nrows[i] > 0L && (nr%%nrows[i] == 0L)) {            xi <- unclass(xi)            fixed <- TRUE            for (j in seq_along(xi)) {                xi1 <- xi[[j]]                if (is.vector(xi1) || is.factor(xi1))                   xi[[j]] <- rep(xi1, length.out = nr)                else if (is.character(xi1) && inherits(xi1, "AsIs"))                   xi[[j]] <- structure(rep(xi1, length.out = nr),                     class = class(xi1))                else if (inherits(xi1, "Date") || inherits(xi1,                   "POSIXct"))                   xi[[j]] <- rep(xi1, length.out = nr)                else {                  fixed <- FALSE                  break                }            }            if (fixed) {                vlist[[i]] <- xi                next            }        }        stop(gettextf("arguments imply differing number of rows: %s",             paste(unique(nrows), collapse = ", ")), domain = NA)    }    value <- unlist(vlist, recursive = FALSE, use.names = FALSE)    vnames <- as.character(unlist(vnames[ncols > 0L]))    if (fix.empty.names && any(noname <- !nzchar(vnames)))         vnames[noname] <- paste0("Var.", seq_along(vnames))[noname]    if (check.names) {        if (fix.empty.names)             vnames <- make.names(vnames, unique = TRUE)        else {            nz <- nzchar(vnames)            vnames[nz] <- make.names(vnames[nz], unique = TRUE)        }    }    names(value) <- vnames    if (!mrn) {        if (length(row.names) == 1L && nr != 1L) {            if (is.character(row.names))                 row.names <- match(row.names, vnames, 0L)            if (length(row.names) != 1L || row.names < 1L ||                 row.names > length(vnames))                 stop("'row.names' should specify one of the variables")            i <- row.names            row.names <- value[[i]]            value <- value[-i]        }        else if (!is.null(row.names) && length(row.names) !=             nr)             stop("row names supplied are of the wrong length")    }    else if (!is.null(row.names) && length(row.names) != nr) {        warning("row names were found from a short variable and have been discarded")        row.names <- NULL    }    class(value) <- "data.frame"    if (is.null(row.names))         attr(value, "row.names") <- .set_row_names(nr)    else {        if (is.object(row.names) || !is.integer(row.names))             row.names <- as.character(row.names)        if (anyNA(row.names))             stop("row names contain missing values")        if (anyDuplicated(row.names))             stop(gettextf("duplicate row.names: %s", paste(unique(row.names[duplicated(row.names)]),                 collapse = ", ")), domain = NA)        row.names(value) <- row.names    }    value})(iri = "http://purl.obolibrary.org/obo/FBbt_00007491", symbol = "",     types = c("Entity", "Adult", "Anatomy", "Class", "Nervous_system",     "Synaptic_neuropil", "Synaptic_neuropil_subdomain"), short_form = "FBbt_00007491",     unique_facets = c("Adult", "Nervous_system", "Synaptic_neuropil_subdomain"    ), label = "fan-shaped body layer 4", check.names = FALSE,     fix.empty.names = TRUE, stringsAsFactors = FALSE): arguments imply differing number of rows: 1, 7, 3

# the same but returning raw JSON
vc2=VfbConnect(raw=TRUE)
vc2$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE)
#> [{'label': 'Kenyon cell', 'symbol': 'KC', 'id': 'FBbt_00003686', 'tags': 'Entity|Anatomy|Cell|Class|Nervous_system|Neuron|has_subClass|hasScRNAseq', 'parents_label': 'anterior ectoderm derivative|mushroom body intrinsic neuron', 'parents_id': 'FBbt_00025991|FBbt_00007484'}]

if (FALSE) { # \dontrun{
# specify a particular neo4j graph database endpoint
vc = VfbConnect(neo_endpoint="http://pdb.p2.virtualflybrain.org")
} # }