Return summary of datasets available from VFB

Return summary of datasets available from VFB

Usage

vfb_datasets()

Value

A data.frame with columns including DOI

Examples

# \donttest{
df=vfb_datasets()
head(df)
#>                                                      label symbol
#> 1                                EM L1 Andrade et al. 2019       
#> 2    MBONs and split-GAL4 lines that target them (Aso2014)       
#> 3 split-GAL4 lines for dopaminergic neurons (AsoRubin2016)       
#> 4                           EM FAFB Baltruschat et al 2021       
#> 5                                EM L1 Barnes et al., 2022       
#> 6                    EM FAFB Bates and Schlegel et al 2020       
#>                  id                                     tags
#> 1       Andrade2019 Entity|DataSet|Individual|has_image|L1EM
#> 2           Aso2014      Entity|DataSet|Individual|has_image
#> 3      AsoRubin2016      Entity|DataSet|Individual|has_image
#> 4   Baltruschat2021 Entity|DataSet|Individual|has_image|FAFB
#> 5        Barnes2022           Entity|DataSet|Individual|L1EM
#> 6 BatesSchlegel2020 Entity|DataSet|Individual|has_image|FAFB
#>                                                                                                  description
#> 1                                                        L1 EM reconstructed neurons from Andrade et al 2019
#> 2                                                                                                       <NA>
#> 3 For comparison of the properties of memories formed by optogenetic activation of individual DAN cell types
#> 4                                                  FAFB EM reconstructed neurons from Baltruschat et al 2021
#> 5                                                       L1 EM reconstructed neurons from Barnes et al., 2022
#> 6                                           FAFB EM reconstructed neurons from Bates and Schlegel et al 2020
#>                                                miniref     FlyBase     PMID
#> 1 Andrade et al., 2019, Curr. Biol. 29(3): 412--425.e3 FBrf0241389 30661802
#> 2                    Aso et al., 2014, eLife 3: e04577 FBrf0227179 25535793
#> 3                 Aso and Rubin, 2016, eLife 5: e16135 FBrf0233230 27441388
#> 4   Baltruschat et al., 2021, Cell Rep. 34(11): 108871 FBrf0248487 33730583
#> 5        Barnes et al., 2022, PLoS ONE 17(4): e0266064 FBrf0253082 35377898
#> 6                      Bates and Schlegel et al., 2020                     
#>                            DOI
#> 1    10.1016/j.cub.2018.12.012
#> 2          10.7554/eLife.04577
#> 3          10.7554/eLife.16135
#> 4 10.1016/j.celrep.2021.108871
#> 5 10.1371/journal.pone.0266064
#> 6    10.1016/j.cub.2020.06.042
subset(df, !nzchar(DOI))
#>                                              label symbol                   id
#> 93        Split-GAL4 lines from Baker et al., 2022              SplitBaker2022
#> 96       Split-GAL4 lines from Cheong et al., 2024             SplitCheong2024
#> 100        Split-GAL4 lines from Feng et al., 2014               SplitFeng2014
#> 111     Split-GAL4 lines from Longden et al., 2023            SplitLongden2023
#> 118        Split-GAL4 lines from Reyn et al., 2017               SplitReyn2017
#> 119       Split-GAL4 lines from Rubin et al., 2024              SplitRubin2024
#> 120 Split-GAL4 lines from Schlichting et al., 2019        SplitSchlichting2019
#> 126    Split-GAL4 lines from Takemura et al., 2017           SplitTakemura2017
#> 130     Split-GAL4 lines from Tuthill et al., 2013            SplitTuthill2013
#> 131     Split-GAL4 lines from Vijayan et al., 2023            SplitVijayan2023
#> 145            L3 Larval CNS Template (Truman2016)                  Truman2016
#> 146              Truman Larval Flip-Out Collection              TrumanWood2018
#> 147              Truman Larval Flip-Out Collection        TrumanWood2018public
#> 155             L3 neuropils (WoodHartenstein2018)         WoodHartenstein2018
#>                                    tags
#> 93            Entity|DataSet|Individual
#> 96            Entity|DataSet|Individual
#> 100           Entity|DataSet|Individual
#> 111           Entity|DataSet|Individual
#> 118           Entity|DataSet|Individual
#> 119           Entity|DataSet|Individual
#> 120           Entity|DataSet|Individual
#> 126           Entity|DataSet|Individual
#> 130           Entity|DataSet|Individual
#> 131           Entity|DataSet|Individual
#> 145 Entity|DataSet|Individual|has_image
#> 146 Entity|DataSet|Individual|has_image
#> 147 Entity|DataSet|Individual|has_image
#> 155 Entity|DataSet|Individual|has_image
#>                                                                                                                                                         description
#> 93                                                                                                                         Split-GAL4 lines from Baker et al., 2022
#> 96                                                                                                                        Split-GAL4 lines from Cheong et al., 2024
#> 100                                                                                                                         Split-GAL4 lines from Feng et al., 2014
#> 111                                                                                                                      Split-GAL4 lines from Longden et al., 2023
#> 118                                                                                                                         Split-GAL4 lines from Reyn et al., 2017
#> 119                                                                                                                        Split-GAL4 lines from Rubin et al., 2024
#> 120                                                                                                                  Split-GAL4 lines from Schlichting et al., 2019
#> 126                                                                                                                     Split-GAL4 lines from Takemura et al., 2017
#> 130                                                                                                                      Split-GAL4 lines from Tuthill et al., 2013
#> 131                                                                                                                      Split-GAL4 lines from Vijayan et al., 2023
#> 145                                                               Template image by James W. Truman and the Fly Light Project Team at HHMI Janelia Research Campus.
#> 146 A selection of neurons from specimens created by Dr. Jim Truman at Janelia Research Campus using the multicolor flip out technique from the FlyLight GMR lines.
#> 147                                                                       Images by James W. Truman and the Fly Light Project Team at HHMI Janelia Research Campus.
#> 155                                                                                                                                                            <NA>
#>     miniref FlyBase PMID DOI
#> 93                          
#> 96                          
#> 100                         
#> 111                         
#> 118                         
#> 119                         
#> 120                         
#> 126                         
#> 130                         
#> 131                         
#> 145                         
#> 146                         
#> 147                         
#> 155                         
# }