Return reference to VfbConnect object

Return reference to VfbConnect object

Usage

VfbConnect(..., raw = FALSE, cache = TRUE)

Arguments

...

Additional arguments passed to Python's `VfbConnect`. See examples.

raw

Whether to return python list containing raw JSON (default FALSE)

cache

Whether to cache the VfbConnect object (default T since this can save some time).

Value

An list containing the processed results or (when raw=TRUE) as python list containing raw JSON.

Details

See https://pypi.org/project/vfb-connect/. The raw=TRUE option is mostly useful to see what output would look like.

Examples

vc=VfbConnect()
# large return, so use str
str(vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686')), max=2)
#> List of 1
#>  $ :List of 10
#>   ..$ term                 :List of 3
#>   ..$ query                : chr "Get JSON for Class"
#>   ..$ version              : chr "c58c844"
#>   ..$ parents              :List of 2
#>   ..$ relationships        :List of 3
#>   ..$ related_individuals  : list()
#>   ..$ xrefs                : list()
#>   ..$ anatomy_channel_image:List of 10
#>   ..$ pub_syn              :List of 6
#>   ..$ def_pubs             :List of 3
# do this to explore
# vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686')) %>% View()
vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE)
#> [[1]]
#> [[1]]$label
#> [1] "Kenyon cell"
#> 
#> [[1]]$symbol
#> [1] "KC"
#> 
#> [[1]]$id
#> [1] "FBbt_00003686"
#> 
#> [[1]]$tags
#> [1] "Entity|Anatomy|Cell|Class|Nervous_system|Neuron|has_subClass|hasScRNAseq"
#> 
#> [[1]]$parents_label
#> [1] "anterior ectoderm derivative|mushroom body intrinsic neuron"
#> 
#> [[1]]$parents_id
#> [1] "FBbt_00025991|FBbt_00007484"
#> 
#> 
# the same but returning raw JSON
vc2=VfbConnect(raw=TRUE)
vc2$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE)
#> [{'label': 'Kenyon cell', 'symbol': 'KC', 'id': 'FBbt_00003686', 'tags': 'Entity|Anatomy|Cell|Class|Nervous_system|Neuron|has_subClass|hasScRNAseq', 'parents_label': 'anterior ectoderm derivative|mushroom body intrinsic neuron', 'parents_id': 'FBbt_00025991|FBbt_00007484'}]

# Information about a dataset linked to a paper
vc$neo_query_wrapper$get_DataSet_TermInfo(list('Ito2013'), summary=TRUE)
#> [[1]]
#> [[1]]$label
#> [1] "Ito lab adult brain lineage clone image set"
#> 
#> [[1]]$symbol
#> [1] ""
#> 
#> [[1]]$id
#> [1] "Ito2013"
#> 
#> [[1]]$tags
#> [1] "Entity|DataSet|Individual|has_image"
#> 
#> [[1]]$description
#> [1] "An exhaustive set of lineage clones covering the adult brain from Kei Ito's  lab."
#> 
#> [[1]]$miniref
#> [1] "Ito et al., 2013, Curr. Biol. 23(8): 644--655"
#> 
#> [[1]]$FlyBase
#> [1] "FBrf0221438"
#> 
#> [[1]]$PMID
#> [1] "23541729"
#> 
#> [[1]]$DOI
#> [1] "10.1016/j.cub.2013.03.015"
#> 
#> 

vc$neo_query_wrapper$get_anatomical_individual_TermInfo(list('VFB_00010001'),  summary=TRUE) %>%
vc_df()
#>          label symbol           id
#> 1 fru-F-500075        VFB_00010001
#>                                                                                                                     tags
#> 1 Entity|Adult|Anatomy|Cell|Expression_pattern_fragment|Individual|Nervous_system|Neuron|VFB|has_image|FlyCircuit|NBLAST
#>                                             parents_label
#> 1 expression pattern fragment|adult SMPpv1 lineage neuron
#>                     parents_id data_source    accession
#> 1 VFBext_0000004|FBbt_00050031  FlyCircuit fru-F-500075
#>                                  templates    dataset license
#> 1 JRC2018Unisex|adult brain template JFRC2 Chiang2010        

# Get all terms relevant to a brain region (all parts and all overlapping cells)
# Query by label supported by default
#
vc$get_terms_by_region('fan-shaped body', summary=TRUE) %>%
  vc_df() %>%
  head(10)
#>                                              label  symbol            id
#> 1                          fan-shaped body layer 4         FBbt_00007491
#> 2                          fan-shaped body layer 5         FBbt_00007492
#> 3                          fan-shaped body layer 3         FBbt_00007490
#> 4                       octopaminergic VPM3 neuron OA-VPM3 FBbt_00110151
#> 5                          fan-shaped body layer 6         FBbt_00007493
#> 6                          fan-shaped body layer 9         FBbt_00047032
#> 7                              adult NPF P2 neuron         FBbt_00048121
#> 8       pupal protocerebral LPP2 Tachykinin neuron         FBbt_00110040
#> 9  adult ellipsoid body extrinsic ring neuron ExR3    ExR3 FBbt_00047035
#> 10                adult ExR1 neuron of DM4 lineage         FBbt_00053035
#>                                                                                       tags
#> 1  Entity|Adult|Anatomy|Class|Nervous_system|Synaptic_neuropil|Synaptic_neuropil_subdomain
#> 2  Entity|Adult|Anatomy|Class|Nervous_system|Synaptic_neuropil|Synaptic_neuropil_subdomain
#> 3  Entity|Adult|Anatomy|Class|Nervous_system|Synaptic_neuropil|Synaptic_neuropil_subdomain
#> 4                     Entity|Adult|Anatomy|Cell|Class|Nervous_system|Neuron|Octopaminergic
#> 5  Entity|Adult|Anatomy|Class|Nervous_system|Synaptic_neuropil|Synaptic_neuropil_subdomain
#> 6  Entity|Adult|Anatomy|Class|Nervous_system|Synaptic_neuropil|Synaptic_neuropil_subdomain
#> 7                                    Entity|Adult|Anatomy|Cell|Class|Nervous_system|Neuron
#> 8                        Entity|Adult|Anatomy|Cell|Class|Nervous_system|Neuron|Peptidergic
#> 9                       Entity|Adult|Anatomy|Cell|Class|Nervous_system|Neuron|Serotonergic
#> 10                                   Entity|Adult|Anatomy|Cell|Class|Nervous_system|Neuron
#>                                                                                                                                                                                                                 parents_label
#> 1                                                                                                                                                                                                       fan-shaped body layer
#> 2                                                                                                                                                                                                       fan-shaped body layer
#> 3                                                                                                                                                                                                       fan-shaped body layer
#> 4  primary interneuron|adult octopaminergic subesophageal VPM neuron|adult mandibular octopaminergic VM neuron|mushroom body octopaminergic neuron|adult mushroom body modulatory input neuron|adult extrinsic nodulus neuron
#> 5                                                                                                                                                                                                       fan-shaped body layer
#> 6                                                                                                                                                                                                       fan-shaped body layer
#> 7                                                                                                                                                                           fan-shaped body intrinsic neuron|adult NPF neuron
#> 8                                                                    pupal Tachykinin neuron of the central nervous system|adult central complex neuron|adult peptidergic neuron|adult CNS neuron|adult neurosecretory neuron
#> 9                                                                                                                               adult ellipsoid body extrinsic ring neuron|adult serotonergic neuron|adult DM6 lineage neuron
#> 10                                                                                                                                                   adult DM4 lineage neuron|adult ellipsoid body extrinsic ring neuron ExR1
#>                                                                             parents_id
#> 1                                                                        FBbt_00040035
#> 2                                                                        FBbt_00040035
#> 3                                                                        FBbt_00040035
#> 4  FBbt_00052517|FBbt_00110148|FBbt_00052534|FBbt_00048137|FBbt_00047957|FBbt_00003662
#> 5                                                                        FBbt_00040035
#> 6                                                                        FBbt_00040035
#> 7                                                          FBbt_00049150|FBbt_00048113
#> 8                FBbt_00110062|FBbt_00049819|FBbt_00058202|FBbt_00059274|FBbt_00052007
#> 9                                            FBbt_00003654|FBbt_00049526|FBbt_00050144
#> 10                                                         FBbt_00050254|FBbt_00003655

if (FALSE) { # \dontrun{
# specify a particular neo4j graph database endpoint
vc = VfbConnect(neo_endpoint="http://pdb.p2.virtualflybrain.org")
} # }