Return reference to VfbConnect object

Return reference to VfbConnect object

VfbConnect(..., raw = FALSE)

Arguments

...
Additional arguments passed to Python's `VfbConnect`. See examples.

raw
Whether to return python list containing raw JSON (default FALSE)

Value

An list containing the processed results or (when raw=TRUE) as python list containing raw JSON.

Details

See https://pypi.org/project/vfb-connect/. The raw=TRUE option is mostly useful to see what output would look like.

Examples

vc=VfbConnect()
# large return, so use str
str(vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686')), max=2)

#> List of 1
#>  $ :List of 9
#>   ..$ term                 :List of 3
#>   ..$ query                : chr "Get JSON for Class"
#>   ..$ version              : chr "44725ae"
#>   ..$ parents              :List of 2
#>   ..$ relationships        :List of 4
#>   ..$ xrefs                : list()
#>   ..$ anatomy_channel_image:List of 5
#>   ..$ pub_syn              :List of 3
#>   ..$ def_pubs             :List of 3

# do this to explore
# vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686')) %>% View()
vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE)

#> [[1]]
#> [[1]]$label
#> [1] "Kenyon cell"
#> 
#> [[1]]$symbol
#> [1] ""
#> 
#> [[1]]$id
#> [1] "FBbt_00003686"
#> 
#> [[1]]$tags
#> [1] "Entity|Anatomy|Nervous_system|Cell|Neuron|Class"
#> 
#> [[1]]$parents_label
#> [1] "supraesophageal ganglion neuron|mushroom body intrinsic neuron"
#> 
#> [[1]]$parents_id
#> [1] "FBbt_00001366|FBbt_00007484"
#> 
#> 

vc$neo_query_wrapper$get_DataSet_TermInfo(list('Ito2013'), summary=TRUE)

#> [[1]]
#> [[1]]$label
#> [1] "Ito lab adult brain lineage clone image set"
#> 
#> [[1]]$symbol
#> [1] ""
#> 
#> [[1]]$id
#> [1] "Ito2013"
#> 
#> [[1]]$tags
#> [1] "Entity|Individual|DataSet"
#> 
#> [[1]]$description
#> [1] "An exhaustive set of lineage clones covering the adult brain from Kei Ito's  lab."
#> 
#> [[1]]$miniref
#> [1] "Ito et al., 2013, Curr. Biol. 23(8): 644--655"
#> 
#> [[1]]$FlyBase
#> [1] "FBrf0221438"
#> 
#> [[1]]$PMID
#> [1] "23541729"
#> 
#> [[1]]$DOI
#> [1] "10.1016/j.cub.2013.03.015"
#> 
#> 

vc$neo_query_wrapper$get_anatomical_individual_TermInfo(list('VFB_00010001'))

#> [[1]]
#> [[1]]$term
#> [[1]]$term$core
#> [[1]]$term$core$iri
#> [1] "http://virtualflybrain.org/reports/VFB_00010001"
#> 
#> [[1]]$term$core$symbol
#> [1] ""
#> 
#> [[1]]$term$core$types
#>  [1] "Entity"         "VFB"            "Adult"          "Anatomy"       
#>  [5] "has_image"      "Cell"           "Individual"     "NBLAST"        
#>  [9] "Nervous_system" "Neuron"        
#> 
#> [[1]]$term$core$label
#> [1] "fru-F-500075"
#> 
#> [[1]]$term$core$short_form
#> [1] "VFB_00010001"
#> 
#> 
#> [[1]]$term$description
#> list()
#> 
#> [[1]]$term$comment
#> [1] "OutAge: Adult 5~15 days"
#> 
#> 
#> [[1]]$query
#> [1] "Get JSON for Individual:Anatomy"
#> 
#> [[1]]$version
#> [1] "44725ae"
#> 
#> [[1]]$dataset_license
#> [[1]]$dataset_license[[1]]
#> [[1]]$dataset_license[[1]]$dataset
#> [[1]]$dataset_license[[1]]$dataset$link
#> [1] "http://flybase.org/reports/FBrf0210580.html"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core
#> [[1]]$dataset_license[[1]]$dataset$core$iri
#> [1] "http://virtualflybrain.org/reports/Chiang2010"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$symbol
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$types
#> [1] "Entity"     "Individual" "DataSet"   
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$label
#> [1] "FlyCircuit 1.0 - single neurons (Chiang2010)"
#> 
#> [[1]]$dataset_license[[1]]$dataset$core$short_form
#> [1] "Chiang2010"
#> 
#> 
#> 
#> [[1]]$dataset_license[[1]]$license
#> [[1]]$dataset_license[[1]]$license$icon
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$license$link
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$license$core
#> [[1]]$dataset_license[[1]]$license$core$iri
#> [1] "http://virtualflybrain.org/reports/VFBlicense_FlyCircuit_License"
#> 
#> [[1]]$dataset_license[[1]]$license$core$symbol
#> [1] ""
#> 
#> [[1]]$dataset_license[[1]]$license$core$types
#> [1] "Entity"     "Individual" "License"   
#> 
#> [[1]]$dataset_license[[1]]$license$core$label
#> [1] "FlyCircuit License"
#> 
#> [[1]]$dataset_license[[1]]$license$core$short_form
#> [1] "VFBlicense_FlyCircuit_License"
#> 
#> 
#> 
#> 
#> 
#> [[1]]$parents
#> [[1]]$parents[[1]]
#> [[1]]$parents[[1]]$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00050031"
#> 
#> [[1]]$parents[[1]]$symbol
#> [1] ""
#> 
#> [[1]]$parents[[1]]$types
#> [1] "Entity"         "Adult"          "Anatomy"        "Nervous_system"
#> [5] "Cell"           "Neuron"         "Class"         
#> 
#> [[1]]$parents[[1]]$label
#> [1] "adult DPMpl1 lineage neuron"
#> 
#> [[1]]$parents[[1]]$short_form
#> [1] "FBbt_00050031"
#> 
#> 
#> 
#> [[1]]$relationships
#> [[1]]$relationships[[1]]
#> [[1]]$relationships[[1]]$relation
#> [[1]]$relationships[[1]]$relation$type
#> [1] "expresses"
#> 
#> [[1]]$relationships[[1]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/RO_0002292"
#> 
#> [[1]]$relationships[[1]]$relation$label
#> [1] "expresses"
#> 
#> 
#> [[1]]$relationships[[1]]$object
#> [[1]]$relationships[[1]]$object$iri
#> [1] "http://flybase.org/reports/FBal0276838"
#> 
#> [[1]]$relationships[[1]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[1]]$object$types
#> [1] "Entity"  "Feature" "Class"  
#> 
#> [[1]]$relationships[[1]]$object$label
#> [1] "Scer\\GAL4[fru.P1.D]"
#> 
#> [[1]]$relationships[[1]]$object$short_form
#> [1] "FBal0276838"
#> 
#> 
#> 
#> [[1]]$relationships[[2]]
#> [[1]]$relationships[[2]]$relation
#> [[1]]$relationships[[2]]$relation$type
#> [1] "overlaps"
#> 
#> [[1]]$relationships[[2]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/RO_0002131"
#> 
#> [[1]]$relationships[[2]]$relation$label
#> [1] "overlaps"
#> 
#> 
#> [[1]]$relationships[[2]]$object
#> [[1]]$relationships[[2]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00007055"
#> 
#> [[1]]$relationships[[2]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[2]]$object$types
#> [1] "Entity"                   "Adult"                   
#> [3] "Anatomy"                  "Nervous_system"          
#> [5] "Synaptic_neuropil"        "Synaptic_neuropil_domain"
#> [7] "Class"                   
#> 
#> [[1]]$relationships[[2]]$object$label
#> [1] "superior medial protocerebrum"
#> 
#> [[1]]$relationships[[2]]$object$short_form
#> [1] "FBbt_00007055"
#> 
#> 
#> 
#> [[1]]$relationships[[3]]
#> [[1]]$relationships[[3]]$relation
#> [[1]]$relationships[[3]]$relation$type
#> [1] "overlaps"
#> 
#> [[1]]$relationships[[3]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/RO_0002131"
#> 
#> [[1]]$relationships[[3]]$relation$label
#> [1] "overlaps"
#> 
#> 
#> [[1]]$relationships[[3]]$object
#> [[1]]$relationships[[3]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00045050"
#> 
#> [[1]]$relationships[[3]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[3]]$object$types
#> [1] "Entity"                   "Adult"                   
#> [3] "Anatomy"                  "Nervous_system"          
#> [5] "Synaptic_neuropil"        "Synaptic_neuropil_domain"
#> [7] "Class"                   
#> 
#> [[1]]$relationships[[3]]$object$label
#> [1] "flange"
#> 
#> [[1]]$relationships[[3]]$object$short_form
#> [1] "FBbt_00045050"
#> 
#> 
#> 
#> [[1]]$relationships[[4]]
#> [[1]]$relationships[[4]]$relation
#> [[1]]$relationships[[4]]$relation$type
#> [1] "part_of"
#> 
#> [[1]]$relationships[[4]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/BFO_0000050"
#> 
#> [[1]]$relationships[[4]]$relation$label
#> [1] "is part of"
#> 
#> 
#> [[1]]$relationships[[4]]$object
#> [[1]]$relationships[[4]]$object$iri
#> [1] "http://virtualflybrain.org/reports/VFBexp_FBal0276838"
#> 
#> [[1]]$relationships[[4]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[4]]$object$types
#> [1] "Entity"             "Class"              "Expression_pattern"
#> [4] "Anatomy"           
#> 
#> [[1]]$relationships[[4]]$object$label
#> [1] "Scer\\GAL4[fru.P1.D] expression pattern"
#> 
#> [[1]]$relationships[[4]]$object$short_form
#> [1] "VFBexp_FBal0276838"
#> 
#> 
#> 
#> [[1]]$relationships[[5]]
#> [[1]]$relationships[[5]]$relation
#> [[1]]$relationships[[5]]$relation$type
#> [1] "part_of"
#> 
#> [[1]]$relationships[[5]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/BFO_0000050"
#> 
#> [[1]]$relationships[[5]]$relation$label
#> [1] "is part of"
#> 
#> 
#> [[1]]$relationships[[5]]$object
#> [[1]]$relationships[[5]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00007011"
#> 
#> [[1]]$relationships[[5]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[5]]$object$types
#> [1] "Entity"  "Class"   "Anatomy"
#> 
#> [[1]]$relationships[[5]]$object$label
#> [1] "female organism"
#> 
#> [[1]]$relationships[[5]]$object$short_form
#> [1] "FBbt_00007011"
#> 
#> 
#> 
#> [[1]]$relationships[[6]]
#> [[1]]$relationships[[6]]$relation
#> [[1]]$relationships[[6]]$relation$type
#> [1] "part_of"
#> 
#> [[1]]$relationships[[6]]$relation$iri
#> [1] "http://purl.obolibrary.org/obo/BFO_0000050"
#> 
#> [[1]]$relationships[[6]]$relation$label
#> [1] "is part of"
#> 
#> 
#> [[1]]$relationships[[6]]$object
#> [[1]]$relationships[[6]]$object$iri
#> [1] "http://purl.obolibrary.org/obo/FBbt_00003624"
#> 
#> [[1]]$relationships[[6]]$object$symbol
#> [1] ""
#> 
#> [[1]]$relationships[[6]]$object$types
#> [1] "Entity"         "Adult"          "Anatomy"        "Nervous_system"
#> [5] "Class"         
#> 
#> [[1]]$relationships[[6]]$object$label
#> [1] "adult brain"
#> 
#> [[1]]$relationships[[6]]$object$short_form
#> [1] "FBbt_00003624"
#> 
#> 
#> 
#> 
#> [[1]]$xrefs
#> [[1]]$xrefs[[1]]
#> [[1]]$xrefs[[1]]$link_base
#> [1] "http://flycircuit.tw/modules.php?name=clearpage&op=detail_table&neuron="
#> 
#> [[1]]$xrefs[[1]]$site
#> [[1]]$xrefs[[1]]$site$iri
#> [1] "http://virtualflybrain.org/reports/FlyCircuit"
#> 
#> [[1]]$xrefs[[1]]$site$symbol
#> [1] ""
#> 
#> [[1]]$xrefs[[1]]$site$types
#> [1] "Entity"     "Individual" "Site"      
#> 
#> [[1]]$xrefs[[1]]$site$label
#> [1] "FlyCircuit 1.0"
#> 
#> [[1]]$xrefs[[1]]$site$short_form
#> [1] "FlyCircuit"
#> 
#> 
#> [[1]]$xrefs[[1]]$is_data_source
#> [1] TRUE
#> 
#> [[1]]$xrefs[[1]]$link_postfix
#> [1] ""
#> 
#> [[1]]$xrefs[[1]]$icon
#> [1] ""
#> 
#> [[1]]$xrefs[[1]]$accession
#> [1] "fru-F-500075"
#> 
#> [[1]]$xrefs[[1]]$link_text
#> [1] "fru-F-500075 on FlyCircuit 1.0"
#> 
#> [[1]]$xrefs[[1]]$homepage
#> [1] "http://flycircuit.tw"
#> 
#> 
#> 
#> [[1]]$channel_image
#> [[1]]$channel_image[[1]]
#> [[1]]$channel_image[[1]]$channel
#> [[1]]$channel_image[[1]]$channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00010001"
#> 
#> [[1]]$channel_image[[1]]$channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[1]]$channel$types
#> [1] "Entity"     "Individual" "VFB"       
#> 
#> [[1]]$channel_image[[1]]$channel$label
#> [1] "fru-F-500075-c"
#> 
#> [[1]]$channel_image[[1]]$channel$short_form
#> [1] "VFBc_00010001"
#> 
#> 
#> [[1]]$channel_image[[1]]$image
#> [[1]]$channel_image[[1]]$image$template_channel
#> [[1]]$channel_image[[1]]$image$template_channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00017894"
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$types
#> [1] "Entity"     "Individual" "Template"   "VFB"       
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$label
#> [1] "JFRC2_template_c"
#> 
#> [[1]]$channel_image[[1]]$image$template_channel$short_form
#> [1] "VFBc_00017894"
#> 
#> 
#> [[1]]$channel_image[[1]]$image$index
#> list()
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy
#> [[1]]$channel_image[[1]]$image$template_anatomy$iri
#> [1] "http://virtualflybrain.org/reports/VFB_00017894"
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$types
#> [1] "Entity"         "VFB"            "Adult"          "Anatomy"       
#> [5] "Nervous_system" "Individual"     "Template"       "has_image"     
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$label
#> [1] "adult brain template JFRC2"
#> 
#> [[1]]$channel_image[[1]]$image$template_anatomy$short_form
#> [1] "VFB_00017894"
#> 
#> 
#> [[1]]$channel_image[[1]]$image$image_folder
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/"
#> 
#> 
#> [[1]]$channel_image[[1]]$imaging_technique
#> [[1]]$channel_image[[1]]$imaging_technique$iri
#> [1] "http://purl.obolibrary.org/obo/FBbi_00000251"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$symbol
#> [1] "Confocal"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$types
#> [1] "Entity" "Class" 
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$label
#> [1] "confocal microscopy"
#> 
#> [[1]]$channel_image[[1]]$imaging_technique$short_form
#> [1] "FBbi_00000251"
#> 
#> 
#> 
#> [[1]]$channel_image[[2]]
#> [[1]]$channel_image[[2]]$channel
#> [[1]]$channel_image[[2]]$channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00010001"
#> 
#> [[1]]$channel_image[[2]]$channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[2]]$channel$types
#> [1] "Entity"     "Individual" "VFB"       
#> 
#> [[1]]$channel_image[[2]]$channel$label
#> [1] "fru-F-500075-c"
#> 
#> [[1]]$channel_image[[2]]$channel$short_form
#> [1] "VFBc_00010001"
#> 
#> 
#> [[1]]$channel_image[[2]]$image
#> [[1]]$channel_image[[2]]$image$template_channel
#> [[1]]$channel_image[[2]]$image$template_channel$iri
#> [1] "http://virtualflybrain.org/reports/VFBc_00101567"
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$types
#> [1] "Entity"     "Individual" "Template"  
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$label
#> [1] "JRC2018Unisex_c"
#> 
#> [[1]]$channel_image[[2]]$image$template_channel$short_form
#> [1] "VFBc_00101567"
#> 
#> 
#> [[1]]$channel_image[[2]]$image$index
#> list()
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy
#> [[1]]$channel_image[[2]]$image$template_anatomy$iri
#> [1] "http://virtualflybrain.org/reports/VFB_00101567"
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$symbol
#> [1] ""
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$types
#> [1] "Entity"         "has_image"      "Adult"          "Anatomy"       
#> [5] "Nervous_system" "Individual"     "Template"      
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$label
#> [1] "JRC2018Unisex"
#> 
#> [[1]]$channel_image[[2]]$image$template_anatomy$short_form
#> [1] "VFB_00101567"
#> 
#> 
#> [[1]]$channel_image[[2]]$image$image_folder
#> [1] "http://www.virtualflybrain.org/data/VFB/i/0001/0001/VFB_00101567/"
#> 
#> 
#> [[1]]$channel_image[[2]]$imaging_technique
#> [[1]]$channel_image[[2]]$imaging_technique$iri
#> [1] "http://purl.obolibrary.org/obo/FBbi_00000251"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$symbol
#> [1] "Confocal"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$types
#> [1] "Entity" "Class" 
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$label
#> [1] "confocal microscopy"
#> 
#> [[1]]$channel_image[[2]]$imaging_technique$short_form
#> [1] "FBbi_00000251"
#> 
#> 
#> 
#> 
#> 

# Get all terms relevant to a brain region (all parts and all overlapping cells$
# Query by label supported by default
vc$get_terms_by_region('fan-shaped body') %>%
  vc_df() %>%
  head(10)

#>                                           label symbol            id
#> 1            horizontal fiber system neuron pb3        FBbt_00007572
#> 2                            pontine neuron P1b        FBbt_00007528
#> 3  fan-shaped body intrinsic unicolumnar neuron        FBbt_00049821
#> 4     fan-shaped body extrinsic columnar neuron        FBbt_00049822
#> 5                           adult NPF P2 neuron        FBbt_00048121
#> 6    adult protocerebral LPP2 Tachykinin neuron        FBbt_00110043
#> 7                          fan-shaped neuron F8        FBbt_00049535
#> 8      PB slice 8-FB-NO3 anterior domain neuron        FBbt_00111402
#> 9              vertical fiber system neuron pb7        FBbt_00007560
#> 10                     tooth of fan-shaped body        FBbt_00049146
#>                                                                                       tags
#> 1                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 2                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 3                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 4                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 5                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 6                 Entity|Peptidergic_neuron|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 7                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 8                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 9                                    Entity|Adult|Anatomy|Nervous_system|Cell|Neuron|Class
#> 10 Entity|Adult|Anatomy|Nervous_system|Synaptic_neuropil|Synaptic_neuropil_subdomain|Class
#>                                                                                                                                 parents_label
#> 1                                                                                                              horizontal fiber system neuron
#> 2                                                                                                bilateral intrinsic neuron|pontine neuron P1
#> 3  small field neuron of the central complex|supraesophageal ganglion neuron|fan-shaped body intrinsic neuron|fan-shaped body columnar neuron
#> 4                  extrinsic neuron|supraesophageal ganglion neuron|small field neuron of the central complex|fan-shaped body columnar neuron
#> 5                                                                                           fan-shaped body intrinsic neuron|adult NPF neuron
#> 6                                                          adult Tachykinin neuron of the central nervous system|adult central complex neuron
#> 7                                                                                                               dorsal fan-shaped body neuron
#> 8                                                                                                            PB-FB-NO3 anterior domain neuron
#> 9                                                                                                                vertical fiber system neuron
#> 10                                                                                                                synaptic neuropil subdomain
#>                                                 parents_id
#> 1                                            FBbt_00003635
#> 2                              FBbt_00003989|FBbt_00007523
#> 3  FBbt_00007514|FBbt_00001366|FBbt_00049150|FBbt_00049820
#> 4  FBbt_00003660|FBbt_00001366|FBbt_00007514|FBbt_00049820
#> 5                              FBbt_00049150|FBbt_00048113
#> 6                              FBbt_00110061|FBbt_00049819
#> 7                                            FBbt_00049537
#> 8                                            FBbt_00111396
#> 9                                            FBbt_00003634
#> 10                                           FBbt_00040006

# the same but returning raw JSON
vc2=VfbConnect(raw=TRUE)
vc2$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE)

#> [{'label': 'Kenyon cell', 'symbol': '', 'id': 'FBbt_00003686', 'tags': 'Entity|Anatomy|Nervous_system|Cell|Neuron|Class', 'parents_label': 'supraesophageal ganglion neuron|mushroom body intrinsic neuron', 'parents_id': 'FBbt_00001366|FBbt_00007484'}]

if (FALSE) {
# specify a particular neo4j graph database endpoint
vc = VfbConnect(neo_endpoint="http://pdb.p2.virtualflybrain.org")
}